Added unsigned int casts to make sure that Urltext tag values won't change after being cast to long or unsigned long on x86-64.
git-svn-id: https://svn.aros.org/svn/aros/trunk@28545 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/mui/classes/urltext/mcc/Urltext_mcc_private.h
@@ -28,7 +28,7 @@
 
 /***********************************************************************/
 
-#define MUISN_Alfie     0xFEC9
+#define MUISN_Alfie     ((unsigned int) 0xFEC9)
 #define TAG_MUI_Alfie   (TAG_USER|(MUISN_Alfie<<16))
 
 /***********************************************************************/
@@ -36,7 +36,7 @@
 /* attributes - Scheme is: [ISGN] */
 enum
 {
-    MUIA_Urltext_Base = (int)(TAG_MUI_Alfie+200),
+    MUIA_Urltext_Base = (unsigned int)(TAG_MUI_Alfie+200),
 
     MUIA_Urltext_MouseOutPen,       /* [IS..] (struct MUI_PenSpec *) PRIVATE!           */
     MUIA_Urltext_MouseOverPen,      /* [IS..] (struct MUI_PenSpec *) PRIVATE!           */
@@ -66,7 +66,7 @@ enum
 /* methods */
 enum
 {
-    MUIAM_Urltext_Base = (int)(TAG_MUI_Alfie+200),
+    MUIAM_Urltext_Base = (unsigned int)(TAG_MUI_Alfie+200),
     MUIM_Urltext_OpenURL,
     MUIM_Urltext_Copy,
     MUIM_Urltext_OpenURLPrefs,
